pkg-py/pyproject.toml
[project]
name = "brand-yaml"
version = "0.1.0"
description = "Read brand yaml files, a unified way to store brand information."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "ruamel-yaml>=0.18.6",
    "pydantic>=2.8.2",
    "jsonschema>=4.23.0",
    "eval-type-backport>=0.2.0",
]

[project.optional-dependencies]
quarto = [
    "pyyaml>=6.0.2",
    "nbformat>=5.10.4",
    "nbclient>=0.10.0",
]
dev = [
    "pyright>=1.1.379",
    "pytest>=8.3.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pyright]
exclude = ["_dev", ".venv", "src/brand_yaml/_brand.py"]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 80
indent-width = 4

target-version = "py39"

[tool.ruff.lint]
select = ['E', 'F', 'W', 'A', 'PLC', 'PLE', 'PLW', 'I']
ignore = ["E501"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"
